pyproject.toml
[project]
name = "engine-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "cython>=3.2.3",
    "pkrbot>=1.1.0",
    "python-dotenv>=1.0.0",
    "openai==0.28.0",
]
